Fix bug #81481 (xml_get_current_byte_index limited to 32-bit numbers on 64-bit builds)
The return value is long in both expat and expat2 (with XML_LARGE_SIZE
not set).

+++ ext/xml/expat_compat.h
@@ -144,7 +144,7 @@ PHP_XML_API int  XML_GetErrorCode(XML_Parser);
 PHP_XML_API const XML_Char *XML_ErrorString(int);
 PHP_XML_API int  XML_GetCurrentLineNumber(XML_Parser);
 PHP_XML_API int  XML_GetCurrentColumnNumber(XML_Parser);
-PHP_XML_API int  XML_GetCurrentByteIndex(XML_Parser);
+PHP_XML_API long XML_GetCurrentByteIndex(XML_Parser);
 PHP_XML_API int  XML_GetCurrentByteCount(XML_Parser);
 PHP_XML_API const XML_Char *XML_ExpatVersion(void);
 PHP_XML_API void XML_ParserFree(XML_Parser);
